Healthbar manaBar display and regen

+++ Assets/Scripts/Player/PlayerStats.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 [CreateAssetMenu (fileName = "New Player", menuName = "Stats/Player")]
 public class PlayerStats : ScriptableObject {
-    public int maxHealth;
-    public int health;
-    public int healthRegen;
-    public int maxMana;
-    public int mana;
-    public int manaRegen;
-    public int armour;
+    public float maxHealth;
+    public float health;
+    public float healthRegen;
+    public float maxMana;
+    public float mana;
+    public float manaRegen;
+    public float armour;
     public float movementSpeed;
     public float jumpForce;
     public float gravityScale;
